% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.GenomicRanges}
\alias{as.GenomicRanges}
\title{Convert a \code{data.frame} representation to \code{GenomicRanges} representation}
\usage{
as.GenomicRanges(x)
}
\arguments{
\item{x}{An input \code{data.frame}.}
}
\value{
\code{GenomicRanges} converted from \code{x}.
}
\description{
Convert a \code{data.frame} representation to \code{GenomicRanges} representation
}
\examples{
dt <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"), 
               use_gr = FALSE, genome = "hs37-1kg")
as.GenomicRanges(dt)
}
