% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{cluster_bed}
\alias{cluster_bed}
\title{Cluster intervals}
\usage{
cluster_bed(x, max_dist = 0)
}
\arguments{
\item{x}{A \code{data.table}}

\item{max_dist}{Maximum distance between features allowed for features to be
merged. Default is 0. That is, overlapping and/or book-ended features are
merged.}
}
\value{
A \code{data.table}. Compared
}
\description{
This operation is similar to \verb{bedtools cluster}. It reports each set of
overlapping or “book-ended” features in an interval file.
}
\examples{
tbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))

# Basic usage
clustered <- cluster_bed(tbl)
head(clustered)

# Change the maximum distance allowed
clustered <- cluster_bed(tbl, max_dist = 10)
head(clustered)
}
\references{
Manual page of \verb{bedtools cluster}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/cluster.html}
}
\seealso{
\code{\link[=merge_bed]{merge_bed()}}
}
