% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{complement_bed}
\alias{complement_bed}
\title{Generate complement}
\usage{
complement_bed(x, genome = NULL)
}
\arguments{
\item{x}{A \code{GRanges}.}

\item{genome}{Specify the reference genome for the BED file. \code{genome} can be
a valid genome name in \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo()}}, e.g. \code{GRCh37}, or
\verb{hs37-1kg}, which is a genome shipped with this package, or any custom
chromosome size files (local or remote). If \code{NULL}, will try to obtain such
information from input data. Refer to \code{\link[=read_bed]{read_bed()}}.}
}
\value{
A \code{GRanges} which is the complement of \code{x}.
}
\description{
This function returns all intervals in a genome that are not covered by at least one interval in the input.
}
\examples{
# Load BED tables
tbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"), use_gr = FALSE)

# Basic usage
result <- complement_bed(tbl, "hs37-1kg")
head(result)
}
\references{
Manual page of \verb{bedtools complement}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/complement.html}
}
\seealso{
\code{\link[=subtract_bed]{subtract_bed()}}
}
