% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2bedtorch}
\alias{df2bedtorch}
\title{Convert a \code{data.frame} to bedtorch data}
\usage{
df2bedtorch(x, genome = NULL, zero_based = TRUE)
}
\arguments{
\item{x}{A \code{data.frame}. The first three columns should be chrom, start and
end. For column names, refer to \code{\link[GenomicRanges:makeGRangesFromDataFrame]{GenomicRanges::makeGRangesFromDataFrame()}}.}

\item{genome}{A character value specifying the genome name.}

\item{zero_based}{If true, the start/end are zero-based, semi-inclusive, which is BED standard.}
}
\value{
A bedtorch data object.
}
\description{
Convert a \code{data.frame} to bedtorch data
}
\examples{
df2bedtorch(data.frame(chrom = "1", start = 1, end = 10), genome = "hs37-1kg")
}
\seealso{
\code{\link[GenomicRanges:makeGRangesFromDataFrame]{GenomicRanges::makeGRangesFromDataFrame()}}
}
