% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.available_colname}
\alias{.available_colname}
\title{Find a new column name which does not conflict with existing names}
\usage{
.available_colname(dt, name)
}
\description{
Find a new column name which does not conflict with existing names
}
\examples{
# Returns "chrom.1"
.available_colname(dt, "chrom")
}
