% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{exclude_bed}
\alias{exclude_bed}
\title{Exclude certain regions}
\usage{
exclude_bed(
  x,
  y,
  max_gap = -1L,
  min_overlap = 0L,
  min_overlap_type = c("bp", "frac1", "frac2")
)
}
\arguments{
\item{x}{A \code{GRanges} object.}

\item{y}{A \code{GRanges} object.}

\item{max_gap}{The largest gap for two intervals to be considered as
overlapping. Default is -1 (no gap allowed, adjacent intervals not allowd).}

\item{min_overlap}{The smallest overlapping region for two intervals to be
considered as overlapping. Default is 0.}

\item{min_overlap_type}{A character value indicating how \code{min_overlap} is
interpreted. \code{bp} means \code{min_overlap} is the number of base pairs. \code{frac1}
means \code{min_overlap} is the minimum overlap required as a fraction of \code{x}.
Similarly, \code{frac2} means \code{min_overlap} is the minimum overlap required as a
fraction of \code{y}. Similar to \verb{bedtools intersect}'s \code{-f} and \code{-F} arguments.}
}
\value{
A \code{GRanges}.
}
\description{
Return original entries in \code{x} that do not overlap with any intervals in \code{y}.
Essentially, this function is just a short-hand for
\code{\link[=intersect_bed]{intersect_bed()}} in \code{exclude} mode.
}
\examples{
# Load BED tables
tbl_x <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"), use_gr = FALSE)
tbl_y <- read_bed(system.file("extdata", "example_intersect_y.bed", package = "bedtorch"), use_gr = FALSE)

# Basic usages
head(exclude_bed(tbl_x, tbl_y))
}
\references{
Manual page of \verb{bedtools intersect}: \url{https://bedtools.readthedocs.io/en/latest/content/tools/intersect.html}
}
\seealso{
\code{\link[=intersect_bed]{intersect_bed()}} \code{\link[GenomicRanges:findOverlaps-methods]{GenomicRanges::findOverlaps()}}
}
