% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_seqinfo}
\alias{get_seqinfo}
\title{Get a \code{GenomeInfoDb::Seqinfo} object}
\usage{
get_seqinfo(genome, genome_name = NULL)
}
\arguments{
\item{genome}{A canonical genome name, e.g. GRCh37, or a custom genome name
recognized by bedtorch, e.g. hs37-1kg, or a path/URL to a chromosome size
file. If \code{NULL}, return \code{NULL}.}

\item{genome_name}{Optional character vector. Only works when \code{genome} is a
file path/URL, and specify the name of the genome. If \code{NULL}, the name is
guessed from the file name.}
}
\description{
Get a \code{GenomeInfoDb::Seqinfo} object
}
