% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{grow_bed}
\alias{grow_bed}
\title{Grow intervals}
\usage{
grow_bed(
  x,
  grow = 1L,
  grow_type = c("bp", "frac"),
  mode = c("both", "left", "right")
)
}
\description{
This function resizes each interval.
}
