% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_seqinfo}
\alias{hs_seqinfo}
\title{hs_seqinfo}
\format{
An object of class \code{list} of length 33.
}
\usage{
hs_seqinfo
}
\description{
Prefetched Homo sapiens seqinfo
}
\keyword{datasets}
