% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{intersect_bed}
\alias{intersect_bed}
\title{Apply \code{intersect} operation between two tables}
\usage{
intersect_bed(
  x,
  y,
  mode = c("default", "exclude", "wa", "wb", "wo", "unique", "loj"),
  max_gap = 0,
  min_overlap = 1
)
}
\arguments{
\item{mode}{\code{default}: Similar to \verb{bedtools intersect}'s default behavior.

\code{wa}: Similar to \verb{bedtools intersect}'s \code{-wa} argument.

\code{wb}: Similar to \verb{bedtools intersect}'s \code{-wb} argument.

\code{unique}: Similar to \verb{bedtools intersect}'s \code{-u} argument.}

\item{max_gap}{The largest gap for two intervals to be considered as
overlapping. Default is 0 (no gap allowed).}

\item{min_overalp}{The smallest overlapping region for two intervals to be
considered as overlapping. Default is 1.}
}
\description{
Apply \code{intersect} operation between two tables
}
