% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{is_gzip}
\alias{is_gzip}
\title{Check if a file name is a gzip file}
\usage{
is_gzip(file_path)
}
\description{
Check if a file name is a gzip file
}
\examples{
is_gzip("foo.gz")
is_gzip("http://foo.bar/example.bed.gz")
is_gzip(c("foo.gz", "bar.gz"))
}
