% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{jaccard_bed}
\alias{jaccard_bed}
\title{Jaccard distance}
\usage{
jaccard_bed(
  x,
  y,
  min_overlap = 1,
  min_overlap_type = c("bp", "frac1", "frac2")
)
}
\arguments{
\item{x}{A \code{GRanges}.}

\item{y}{A \code{GRanges}.}

\item{min_overlap}{The smallest overlapping region for two intervals to be
considered as overlapping. Default is 1.}

\item{min_overlap_type}{A character value indicating how \code{min_overlap} is
interpreted. \code{bp} means \code{min_overlap} is the number of base pairs. \code{frac1}
means \code{min_overlap} is the minimum overlap required as a fraction of \code{x}.
Similarly, \code{frac2} means \code{min_overlap} is the minimum overlap required as a
fraction of \code{y}. Similar to \verb{bedtools intersect}'s \code{-f} and \code{-F} arguments.}
}
\value{
The Jaccard statistics.
}
\description{
Calculate Jaccard distance between two BED-like data sets
}
\examples{
# Load BED tables
x <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"), use_gr = FALSE)
y <- read_bed(system.file("extdata", "example_intersect_y.bed", package = "bedtorch"), use_gr = FALSE)

# Basic usages
result <- jaccard_bed(x, y) 
head(result)
}
\references{
Manual page of \verb{bedtools jaccard}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/jaccard.html}
}
\seealso{
\code{\link[=intersect_bed]{intersect_bed()}}
}
