% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_bed}
\alias{load_bed}
\alias{load_fragments}
\title{Load fragment data files or BED data files}
\usage{
load_bed(file_path, region = NULL)

load_fragments(file_path, region = NULL)
}
\arguments{
\item{file_path}{Path to the data file (in tab-separated values format, a.k.a "tsv").}

\item{region}{Which genomic region to load. Can be either a character vector
following the UCSC convention, e.g. \code{21:1-9411192}, or a
GenomicRanges::GRanges object. Default is \code{NULL}.}
}
\value{
A \code{data.table} object
}
\description{
Fragment data files are essentially BED data files. \code{load_fragments} and
\code{load_bed} will load these data files into \code{data.table} objects.
}
\details{
If \code{regions} is not \code{NULL}, \code{load_fragments} will invoke \code{tabix} to load the data.
}
