% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_genbed}
\alias{load_genbed}
\title{Load a data file in GENBED format}
\usage{
load_genbed(
  file_path,
  region = NULL,
  col_names = NULL,
  na = c("", "NA"),
  col_types = NULL,
  ...
)
}
\arguments{
\item{file_path}{Path or URL to the data file. Can be a local path, file://,
http://, https://, ftp:// and sftp://}

\item{col_names}{can be NULL (infer from file), or a character vector
(user-provided column names). In the infer-from-file scenario, if there is
a column header, we will have the column names. However, if the column header
is missing, then we will have the following column names: chrom, start, end,
X4, X5, ...}

\item{na}{Character vector of strings to interpret as missing values.}

\item{col_types}{A character string represents column names. Refer to
\code{col_type} in readr. If NULL, then the types for the first three columns
are character, integer, integer, and the rest are guessed from the data.}
}
\description{
This function loads a genbed file, which is simply a BED-like tsv file,
with a header line (optional) containing column names, and the types of
the first three columns are guaranteed to be "cii"
}
