% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_genbed_tabix}
\alias{load_genbed_tabix}
\title{Load GENBED using tabix}
\usage{
load_genbed_tabix(
  file_path,
  region,
  col_names = NULL,
  na = c("", "NA"),
  col_types = NULL,
  ...
)
}
\arguments{
\item{file_path}{Path to the data file. Should be compressed by BGZIP,
locally present, and come along with a paired tabix index file}
}
\description{
Load GENBED using tabix
}
