% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{make_random_bed}
\alias{make_random_bed}
\title{Randomly place intervals of fixed width across genome}
\usage{
make_random_bed(
  n,
  interval_width,
  chrom_sizes = c("hg19", "hg38"),
  chrom = NULL,
  seed = NULL,
  sort = TRUE
)
}
\arguments{
\item{n}{Number of intervals to generate.}

\item{interval_width}{An integer value of the width of the interval.}

\item{chrom_sizes}{Provide the chromosome sizes data. If \code{hg19} or \code{hg38},
the embedded chromosome sizes for hg19 or hg38 will be used. Otherwise, it
need to be the path of a user-provided file.}

\item{chrom}{A character vector. If provided, only generate random intervals
for the specified chromosome(s).}

\item{seed}{An integer seed for the random number generator. If \code{NULL}, the
seed is randomly chosen.}
}
\value{
A \code{data.table} of generated intervals.
}
\description{
This function is aware of chromosome size differences and intervals are
randomly generated in proportion to chromosome sizes. In another word, larger
chromosomes will have more randomly generated intervals.
}
\examples{
# Generate 100 500kbp-intervals for chr20 and chr22, hg38
result <- make_random_bed(100, 500e3L, chrom_sizes = "hg38", chrom = c("chr20", "chr22"))
head(result)

# Generate 100 500bp unsorted intervals for all chromosomes of hg19 (default reference genome)
result <- make_random_bed(100, interval_width = 500, chrom_sizes = "hg19", sort = FALSE)
head(result)
}
\references{
Manual page for \verb{bedtools random}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/random.html}
}
