% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{min_overlap_filter}
\alias{min_overlap_filter}
\title{Filter \code{overlap_table} by \code{min_overlap}}
\usage{
min_overlap_filter(x, y, overlap_table, min_overlap, min_overlap_type)
}
\value{
A new \code{overlap_table} containing only the records pass the filter
}
\description{
Filter \code{overlap_table} by \code{min_overlap}
}
