% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncbi_seqinfo}
\alias{ncbi_seqinfo}
\title{ncbi_seqinfo}
\format{
An object of class \code{list} of length 22.
}
\usage{
ncbi_seqinfo
}
\description{
Prefetched NCBI seqinfo
}
\keyword{datasets}
