% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_bedtorch_table}
\alias{new_bedtorch_table}
\title{Constructor of bedtorch_table}
\usage{
new_bedtorch_table(dt, genome = NULL)
}
\arguments{
\item{genome}{A character value specifying the genome name, or a
\code{GenomeInfoDb::Seqinfo} object.}
}
\description{
Constructor of bedtorch_table
}
