% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{overlapping_intervals}
\alias{overlapping_intervals}
\title{For each interval pair, get the overlapping start and end}
\usage{
overlapping_intervals(x, y, overlap_table)
}
\description{
For each interval pair, get the overlapping start and end
}
