% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{post_process_table}
\alias{post_process_table}
\title{Set \code{dt}'s column names, types, and set the index}
\usage{
post_process_table(dt)
}
\description{
Set \code{dt}'s column names, types, and set the index
}
\details{
Modifies \code{dt} in-place.
}
