% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_bed}
\alias{read_bed}
\title{Load a BED-format file}
\usage{
read_bed(file_path, range = NULL, ...)
}
\arguments{
\item{file_path}{Path to the data file. It can be either a local file, or a remote URL.}

\item{range}{A genomic range character vector. Must follow standard genomic
range notation format, e.g. chr1:1001-2000}

\item{...}{Other arguments to be passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\description{
This function loads the input file as a \code{data.table} object. The file can be
either local or remote, and can be either plain text or gzip-compressed.
Furthermore, this function supports range-loading by providing a genomic
range in the following syntax: "chr1:1-100".
}
\details{
Note: for loading remote data files, currently this function depends on
tabix.c 0.2.5, which doesn't not support HTTPS protocol. In the next step, I
plan to turn to htslib, and the this function can load remote data files
through HTTPS.
}
\examples{
bedtbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))
bedtbl <- read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
                   range = "1:3001-4000")

# Does not work. Currently HTTPS is not supported
# bedtbl <- read_bed("https://yourdomain.com/example-02.bedGraph.gz", 
#                    range = "1:1000-2000")
}
\seealso{
\code{\link[data.table:fread]{data.table::fread()}}
}
