% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_metadata}
\alias{read_metadata}
\title{metadata are defined as # comment lines, in the form of #key=value}
\usage{
read_metadata(file_path)
}
\description{
metadata are defined as # comment lines, in the form of #key=value
}
