% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_table}
\alias{read_table}
\title{Load a BED-format file}
\usage{
read_table(file_path, range = NULL, ...)
}
\arguments{
\item{range}{A genomic range character vector. Must be}
}
\description{
This function loads the input file as a \code{data.table} object. The file can be
either local or remote, and can be either plain text or gzip-compressed.
Furthermore, this function supports range-loading by providing a genomic
range in the following syntax: "chr1:1-100".
}
