% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{reldist_bed}
\alias{reldist_bed}
\title{Relative distance distribution}
\usage{
reldist_bed(x, y)
}
\arguments{
\item{x}{A \code{GRanges}.}

\item{y}{A \code{GRanges}.}
}
\value{
The relative distance distribution.
}
\description{
Calculate the relative distance distribution between two BED data tables.
}
\examples{
# Load BED tables
x <- read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"), use_gr = FALSE)
y <- read_bed(system.file("extdata", "example2_window.bed", package = "bedtorch"), use_gr = FALSE)

# Basic usages
reldist_bed(x, y) 
}
\references{
Manual page of \verb{bedtools reldist}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/reldist.html}

Exploring Massive, Genome Scale Datasets with the GenometriCorr Package.
Favorov A, Mularoni L, Cope LM, Medvedeva Y, Mironov AA, et al. (2012) PLoS
Comput Biol 8(5): e1002529.
\href{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1002529}{doi:10.1371/journal.pcbi.1002529}
}
