% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rollmean}
\alias{rollmean}
\title{Calculate rolling means}
\usage{
rollmean(
  x,
  k,
  na_pad = FALSE,
  na.rm = FALSE,
  align = c("center", "left", "right")
)
}
\arguments{
\item{x}{A numeric vector.}

\item{k}{integer width of the rolling window.}

\item{na_pad}{Logical value about whether to pad left/right ends with NAs.}

\item{na.rm}{Similar to \code{na.rm} in \code{\link[base:mean]{base::mean()}}, a logical value indicating
whether NA values should be stripped before the computation proceeds.}

\item{align}{Character specifying whether the index of the result should be
left- or right-aligned or centered (default) compared to the rolling window
of observations.}
}
\value{
The rolling mean.
}
\description{
Compared with \code{\link[zoo:rollmean]{zoo::rollmean()}}, this function is implemented in C and is one
magnitude faster.
}
\examples{
x <- 1:10

# Basic usage
rollmean(x, k = 3)

# Align at the right end, and use NA for padding
rollmean(x, k = 3, na_pad = TRUE, align = "right")
}
\seealso{
\code{\link[zoo:rollmean]{zoo::rollmean()}}
}
