% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{shuffle_bed}
\alias{shuffle_bed}
\title{Shuffle intervals}
\usage{
shuffle_bed(x, chrom_sizes = c("hg19", "hg38"), excluded_region = NULL)
}
\arguments{
\item{x}{A \code{data.table}.}

\item{chrom_sizes}{Provide the chromosome sizes data. If \code{hg19} or \code{hg38},
the embedded chromosome sizes for hg19 or hg38 will be used. Otherwise, it
need to be the path of a user-provided file.}

\item{excluded_region}{A \code{data.table} containing regions that you don't want
to exclude from the shuffle. Shuffled intervals are guaranteed not to fall
within such regions.}
}
\value{
A \code{data.table} containing shuffled features.
}
\description{
This function randomly permutes genomic locations among a genome defined in a
genome file. Note: the permutation stays in the same chromosome.
Cross-chromosome shuffle is not supported at this point.
}
\examples{
# Load BED tables
tbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))
excluded <- read_bed(system.file("extdata", "example_intersect_y.bed", package = "bedtorch"))

# Basic usage
result <- shuffle_bed(tbl, chrom_sizes = "hg19", excluded_region = excluded)
}
\references{
Manual page of \verb{bedtools shuffle}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/shuffle.html}
}
