% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{slop_bed}
\alias{slop_bed}
\title{Slop intervals}
\usage{
slop_bed(
  x,
  chrom_sizes = c("hg19", "hg38"),
  slop = 1L,
  slop_type = c("bp", "frac"),
  mode = c("both", "left", "right")
)
}
\arguments{
\item{x}{A \code{GRanges} object.}

\item{chrom_sizes}{Provide the chromosome sizes data. If \code{hg19} or \code{hg38},
the embedded chromosome sizes for hg19 or hg38 will be used. Otherwise, it
need to be the path of a user-provided file. If \code{NULL}, the resizing will
not be restricted to the size of the chromosome.}

\item{slop}{The number of base pairs to slop, or the percentage of the
interval's width (refer to \verb{bedtools slop}'s \code{-pct} argument).}

\item{slop_type}{If \code{bp}, \code{slop} is an integer indicating the number of base
pairs. If \code{frac}, \code{slop} is a float point number indicating the percentage.}

\item{mode}{If \code{both}, the resizing takes place on both ends of the interval.
Otherwise, the resizing is only for the \code{left} end or the \code{right} end.}
}
\value{
A slopped \code{data.tabe} object.
}
\description{
Similar to \verb{bedtools slop}, this function increases the size of each feature
in a feature file by a user-defined number of bases. Will restrict the
resizing to the size of the chromosome.
}
\examples{
# Load data
tbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"), use_gr = FALSE)

# Slop by 10 on both ends
result <- slop_bed(tbl, slop = 10L, slop_type = "bp", mode = "both")

# Slop by 10 only on the right end
result <- slop_bed(tbl, slop = 10L, slop_type = "bp", mode = "right")
head(result)

# Slop by 20\% on the left end
result <- slop_bed(tbl, slop = 0.2, slop_type = "frac", mode = "left")
head(result)
}
\references{
Manual page of \verb{bedtools slop}: \url{https://bedtools.readthedocs.io/en/latest/content/tools/slop.html}
}
