% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{subtract_bed}
\alias{subtract_bed}
\title{Subtract}
\usage{
subtract_bed(
  x,
  y,
  min_overlap = 1,
  min_overlap_type = c("bp", "frac1", "frac2")
)
}
\arguments{
\item{x}{A \code{data.table} object.}

\item{y}{A \code{data.table} object.}

\item{min_overlap}{See \code{\link[=intersect_bed]{intersect_bed()}}.}

\item{min_overlap_type}{\code{\link[=intersect_bed]{intersect_bed()}}.}
}
\value{
A \code{data.table} after subtracting \code{y} from \code{x}
}
\description{
This function searches for intervals in \code{y} that overlaps with \code{x}, and
remove the overlapping portion of \code{x}.
}
\examples{
# Load BED tables
tbl_x <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))
tbl_y <- read_bed(system.file("extdata", "example_intersect_y.bed", package = "bedtorch"))

# Basic usage
result <- subtract_bed(tbl_x, tbl_y)
head(result)

# Perform the mapping, requiring the minimum overlapping of 5bp
result <- subtract_bed(tbl_x, tbl_y, min_overlap = 5, min_overlap_type = "bp")
head(result)
}
\references{
Manual page of \verb{bedtools subtract}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/subtract.html}
}
\seealso{
\code{\link[=intersect_bed]{intersect_bed()}}
}
