% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_bed}
\alias{write_bed}
\title{Write a \code{data.table} to file}
\usage{
write_bed(x, file_path, tabix_index = TRUE, batch_size = NULL, ...)
}
\arguments{
\item{x}{A \code{data.table} object to write.}

\item{file_path}{Path of the output file. If the extension name is \code{.gz} or
\code{.bgz}, the output will be compressed in BGZIP format.}

\item{tabix_index}{If \code{TRUE}, and \code{file_path} indicates a gzip file, will
create create the tabix index file.}

\item{...}{Other arguments passed to methods. Compliant with \code{data.table::fwrite}.}
}
\description{
Can be a plain text file or a BGZIP file.
}
\examples{
bedtbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))

# Write data to uncompressed file
write_bed(bedtbl,  tempfile(fileext = ".bed"))

# Write data to file and create tabix index
write_bed(bedtbl, tempfile(fileext = ".bed.gz"), tabix_index = TRUE)
}
