% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_table}
\alias{write_table}
\title{Write a \code{data.table} to file}
\usage{
write_table(x, file_path, tabix_index = TRUE, ...)
}
\arguments{
\item{file_path}{Path of the output file. If the extension name is \code{.gz} or
\code{.bgz}, the output will be compressed in BGZIP format.}

\item{tabix_index}{If \code{TRUE}, and \code{file_path} indicates a gzip file, will
create create the tabix index file.}

\item{...}{Other arguments passed to methods. Compliant with \code{data.table::fwrite}.}
}
\description{
Can be a plain text file or a BGZIP file.
}
\examples{
write_tabix_table(data, "example.bed")
write_tabix_table(data, "example.bed.gz", tabix_index = TRUE)
}
