% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.bedtorch_table}
\alias{as.bedtorch_table}
\title{Convert a \code{GenomicRanges} representation to \code{data.table} representation}
\usage{
as.bedtorch_table(gr)
}
\description{
Convert a \code{GenomicRanges} representation to \code{data.table} representation
}
