% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.GenomicRanges}
\alias{as.GenomicRanges}
\title{Convert a \code{data.table} representation to \code{GenomicRanges} representation}
\usage{
as.GenomicRanges(x)
}
\arguments{
\item{x}{An input \code{data.table}.}
}
\value{
\code{GenomicRanges} converted from \code{x}.
}
\description{
Convert a \code{data.table} representation to \code{GenomicRanges} representation
}
