% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.GenomicRanges}
\alias{as.GenomicRanges}
\title{Convert a \code{data.table} representation to \code{GenomicRanges} representation}
\usage{
as.GenomicRanges(dt)
}
\arguments{
\item{genome}{A character value specifying the genome name, or a
\code{GenomeInfoDb::Seqinfo} object.}
}
\description{
Convert a \code{data.table} representation to \code{GenomicRanges} representation
}
