% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{complement_bed}
\alias{complement_bed}
\title{Generate complement}
\usage{
complement_bed(x, chrom_sizes = c("hg19", "hg38"))
}
\arguments{
\item{x}{A \code{GRanges}.}

\item{chrom_sizes}{Provide the chromosome sizes data. If \code{hg19} or \code{hg38},
the embedded chromosome sizes for hg19 or hg38 will be used. Otherwise, it
need to be the path of a user-provided file.}
}
\value{
A \code{GRanges} which is the complement of \code{x}.
}
\description{
This function returns all intervals in a genome that are not covered by at least one interval in the input.
}
\examples{
# Load BED tables
tbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"), use_gr = FALSE)

# Basic usage
result <- complement_bed(tbl, "hg19")
head(result)
}
\references{
Manual page of \verb{bedtools complement}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/complement.html}
}
\seealso{
\code{\link[=subtract_bed]{subtract_bed()}}
}
