% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{make_random_bed}
\alias{make_random_bed}
\title{Randomly place intervals of fixed width across genome}
\usage{
make_random_bed(
  n,
  interval_width,
  genome,
  chrom = NULL,
  seed = NULL,
  sort = TRUE
)
}
\arguments{
\item{n}{Number of intervals to generate.}

\item{interval_width}{An integer value of the width of the interval.}

\item{genome}{Specify the reference genome for the BED file. \code{genome} can be
a valid genome name in \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo()}}, e.g. \code{GRCh37}, or
\verb{hs37-1kg}, which is a genome shipped with this package, or any custom
chromosome size files (local or remote). If \code{NULL}, will try to obtain such
information from input data. Refer to \code{\link[=read_bed]{read_bed()}}.}

\item{chrom}{A character vector. If provided, only generate random intervals
for the specified chromosome(s).}

\item{seed}{An integer seed for the random number generator. If \code{NULL}, the
seed is randomly chosen.}

\item{sort}{Logical value. Should the random BED data be sorted?}
}
\value{
A \code{GRanges} of generated intervals.
}
\description{
This function is aware of chromosome size differences and intervals are
randomly generated in proportion to chromosome sizes. In another word, larger
chromosomes will have more randomly generated intervals.
}
\examples{
# Generate 100 500kbp-intervals for chr20 and chr22, GRCh38
result <- make_random_bed(100, 500e3L, genome = "GRCh38", chrom = c("20", "22"))
head(result)

# Generate 100 500bp unsorted intervals for all chromosomes of GRCh37 (default reference genome)
result <- make_random_bed(100, interval_width = 500, genome = "GRCh37", sort = FALSE)
head(result)
}
\references{
Manual page for \verb{bedtools random}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/random.html}
}
