% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{make_windows}
\alias{make_windows}
\title{Make windows along genome}
\usage{
make_windows(window_size, genome = NULL, chrom = NULL, chrom_sizes = NULL)
}
\arguments{
\item{window_size}{An integer value. Size of the windows in base pairs.}

\item{genome}{Provide the genome identifier, e.g. \code{GRCh37}. Refer to
\code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo()}}.}

\item{chrom}{A character vector. If provided, only generate windows for the
specified chromosome(s).}
}
\value{
A \code{GRanges} object containing the generated windows.
}
\description{
Make windows with the size of \code{window_size} along genome. Must provide
\code{genome} so that intervals are restricted by chromosome sizes.
}
\examples{
# Generate 500kbp intervals for chr20 and chr22, hg38
result <- make_windows(window_size = 500e3L, genome = "hg38", chrom = c("chr20", "chr22"))
head(result)

# Generate 5Mbp intervals for all chromosomes of GRCh37 (default reference genome)
result <- make_windows(window_size = 5e6L, genome = "GRCh37")
head(result)
}
