% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_bed}
\alias{read_bed}
\title{Load a BED-format file}
\usage{
read_bed(
  file_path,
  range = NULL,
  compression = c("detect", "bgzip", "text", "other"),
  tabix_index = NULL,
  download_index = FALSE,
  genome = NULL,
  use_gr = TRUE,
  sep = "\\t",
  ...
)
}
\arguments{
\item{file_path}{Path to the data file. It can be either a local file, or a remote URL.}

\item{range}{A genomic range character vector. Must follow standard genomic
range notation format, e.g. chr1:1001-2000}

\item{compression}{Indicate the compression type. If \code{detect}, this function
will try to guess from \code{file_path}.}

\item{tabix_index}{A character value indicating the location of the tabix
index file. Can be either local or remote. If \code{NULL}, it will be derived
from \code{file_path}.}

\item{download_index}{Whether to download (cache) the tabix index at current
directory.}

\item{genome}{Specify the reference genome for the BED file. \code{genome} can be
a valid genome name in \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo()}}, e.g. \code{GRCh37}, or
\verb{hs37-1kg}, which is a genome shipped with this package, or any custom
chromosome size files (local or remote). Here is a good resource for such
files: \url{https://github.com/igvteam/igv/tree/master/genomes/sizes}.}

\item{use_gr}{If \code{TRUE}, will read the data as a \code{GenomicRanges} object,
otherwise a \code{data.table} object. Generally, we recommend using
\code{GenomicRanges}.}

\item{sep}{The separator between columns. By default, BED files are
tab-delimited, and \code{sep} should be \verb{\\t}. However, sometimes you will
encounter non-standard table files. In such cases, you need to specify the
separator. If \code{auto}, \code{read_bed} will try to guess the separator. For more
details, refer to \code{\link[data.table:fread]{data.table::fread()}}.}

\item{...}{Other arguments to be passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\description{
This function loads the input file as a \code{data.table} object. The file can be
either local or remote, and can be either plain text or gzip-compressed.
Furthermore, this function supports range-loading by providing a genomic
range in the following syntax: "chr1:1-100".
}
\details{
Note: for loading remote data files, currently this function depends on
tabix.c 0.2.5, which doesn't not support HTTPS protocol. In the next step, I
plan to turn to htslib, and the this function can load remote data files
through HTTPS.
}
\examples{
bedtbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"))
head(bedtbl)

# Basic usage
bedtbl <- read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
                  range = "1:3001-4000")
head(bedtbl)

# Specify the reference genome
head(read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
              range = "1:3001-4000",
              genome = "hs37-1kg"))

head(read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
              range = "1:3001-4000",
              genome = "GRCh37"))

head(read_bed(system.file("extdata", "example2.bed.gz", package = "bedtorch"),
              range = "1:3001-4000",
              genome = "https://raw.githubusercontent.com/igvteam/igv/master/genomes/sizes/1kg_v37.chrom.sizes"))

# Load remote BGZIP files with tabix index specified
head(read_bed("https://git.io/JYATB", range = "22:20000001-30000001", tabix_index = "https://git.io/JYAkT"))
}
\seealso{
\code{\link[data.table:fread]{data.table::fread()}}
}
