% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{shuffle_bed}
\alias{shuffle_bed}
\title{Shuffle intervals}
\usage{
shuffle_bed(x, genome = NULL, excluded_region = NULL, sort = TRUE, seed = NULL)
}
\arguments{
\item{x}{A \code{GRanges}.}

\item{genome}{Specify the reference genome for the BED file. \code{genome} can be
a valid genome name in \code{\link[GenomeInfoDb:Seqinfo-class]{GenomeInfoDb::Seqinfo()}}, e.g. \code{GRCh37}, or
\verb{hs37-1kg}, which is a genome shipped with this package, or any custom
chromosome size files (local or remote). If \code{NULL}, will try to obtain such
information from input data. Refer to \code{\link[=read_bed]{read_bed()}}.}

\item{excluded_region}{A \code{GRanges} containing regions that you don't want
to exclude from the shuffle. Shuffled intervals are guaranteed not to fall
within such regions.}

\item{sort}{Logical value indicating whether to sort the shuffled features.
When the input data set is large, sorting can be very expensive. If you
don't need the output to be sorted, set \code{sort} to \code{FALSE}.}

\item{seed}{An integer seed for the random number generator. If \code{NULL}, the
seed is randomly chosen.}
}
\value{
A \code{GRanges} containing shuffled features.
}
\description{
This function randomly permutes genomic locations among a genome defined in a
genome file. Note: the permutation stays in the same chromosome.
Cross-chromosome shuffle is not supported at this point.
}
\examples{
# Load BED tables
tbl <- read_bed(system.file("extdata", "example_merge.bed", package = "bedtorch"), use_gr = FALSE)
excluded <- read_bed(system.file("extdata", "example_intersect_y.bed", package = "bedtorch"), use_gr = FALSE)

# Basic usage
result <- shuffle_bed(tbl)
head(result)

# Shuffle the data, but exclude certain regions. Plus, set the RNG seed to 1
result <- shuffle_bed(tbl, chrom_sizes = "hg19", excluded_region = excluded, seed = 1)
head(result)
}
\references{
Manual page of \verb{bedtools shuffle}:
\url{https://bedtools.readthedocs.io/en/latest/content/tools/shuffle.html}
}
